refactored fetch function to a util file and fixed setloaded issued.
