+++ src/components/NavBar.css
@@ -17,10 +17,9 @@ link.App-header:visited {
     align-items: center;
     justify-content: space-around;
     font-size: calc(10px + 2vmin);
-    color: whitesmoke;
+    color: darkred;
 
 }
-
 
 nav {
     display: flex;
