+++ src/components/Top10Keywords.js
@@ -10,6 +10,8 @@ import jobsContext from '../context/jobsContext.js'
 import sortSlice10 from '../utils/SortSlice.js'
 import Top10ResultLinkEntry from './Top10ResultLinkEntry'
 import { CircularProgress } from '@material-ui/core'
+import FetchJobsData from '../utils/FetchJobsData.js'
+
 
 function Top10Keywords() {
     const [isloaded, setloaded] = useState(false)
@@ -17,34 +19,15 @@ function Top10Keywords() {
 
     useEffect(() => {
         setKeywords([])
-        Keywords.forEach((word) => {
-            const url = `https://data.usajobs.gov/api/search?Keyword=${word}&ResultsPerPage=1&Fields=min${filter ? "&" + filter : ""}`;
-            fetch(url, {
-                method: 'GET',
-                headers: { 
-                  'Content-Type': 'application/json',
-                  'User-Agent': 'gordon.deng.1@us.af.mil',
-                  'Host': 'data.usajobs.gov',
-                  'Authorization-Key': '8anoUcMouBmDxVIpesVxLsxa0z2IyHXh2bL7iHZdvOA='
-                }
-            })
-                .then(response => response.json())
-                .then(data => {
-                    setKeywords((previous) => {
-                        const newKeywords = [...previous];
-                        newKeywords.push({ "word": word, "count": data.SearchResult.SearchResultCountAll });
-                        return newKeywords
-                    })
-                })
-                .then(data => {setloaded(true)})
-        })
+        setloaded(false)
+        FetchJobsData(Keywords, "https://data.usajobs.gov/api/search?Keyword=", setKeywords, filter, setloaded) 
     }, [filter, setKeywords])
 
     const loadingBar = isloaded ?
         (
             <ol>
                 {sortSlice10(keywords).map(obj => {
-                    return <Top10ResultLinkEntry type={'k'} name={obj.word} count={obj.count} />
+                    return <Top10ResultLinkEntry key={obj.jobsId} type={'k'} name={obj.name} count={obj.count} />
                 })}
             </ol>
         ) :
