--- src/utils/FormatJSON.js
@@ -1,7 +1,0 @@
-
-
-function formatJSON (object) {
-    return 
-}
-
-export default formatJSON;
